Refactor scripts

+++ neovim/neovim.sh
@@ -1,12 +1,25 @@
-#!/bin/bash
+#!/usr/bin/env bash
 set -euxo pipefail
 
-# remove vim
-sudo apt purge vim-tiny vim-runtime vim-common -y
+export OS=${OSTYPE:-'linux-gnu'}
+export OS_TYPE=`echo ${OS} | tr -d "[:digit:]"`
 
-# install neovim
-sudo apt update
-sudo apt install -y neovim
+if [ "${OS_TYPE}" == "linux-gnu" ]; then
+    if [ ! -d ~/.config/nvim ]; then
+        if [ -f /etc/debian_version ]; then
+            # remove vim
+            sudo apt purge -y vim-tiny vim-runtime vim-common
+            # install neovim
+            sudo apt update
+            sudo apt install -y neovim
+        fi
+    fi
+elif [ "${OS_TYPE}" == "darwin" ]; then
+    if [ ! -d ~/.config/nvim ]; then
+        # install neovim
+        which brew > /dev/null && brew install neovim
+    fi
+fi
 
 # install neovim configuration file
 CWD=$(dirname $0)
@@ -14,4 +27,4 @@ mkdir -p ~/.config/nvim
 cp $CWD/init.vim ~/.config/nvim/init.vim
 
 # install plugins
-nvim +PlugInstall
+nvim +PlugInstall +qall
